refactor(core): split queuing ports logic into multiple module
